+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Display "Follow" button to guests.
+- Make background color of public About page customizable.
 
 ### Changed